third_party/WebKit/LayoutTests/harness-tests/console_logging.html

<!DOCTYPE html>
<title>Console messages in non-WPT tests should appear in the test output.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

console.log("Single line message");
console.log("Multi-line\nmessage");
test(function() {
  assert_true(false);
}, 'Test needs to fail so that an expected.txt file is written out and be examined for console messages.');

</script>
