third_party/WebKit/LayoutTests/harness-tests/console_logging-expected.txt

CONSOLE MESSAGE: line 7: Single line message
CONSOLE MESSAGE: line 8: Multi-line
message
This is a testharness.js-based test.
FAIL Test needs to fail so that an expected.txt file is written out and be examined for console messages. assert_true: expected true got false
Harness: the test ran to completion.
